+++ R/weather.R
@@ -6,10 +6,10 @@
 #' The information comes from the nearest available weather station and is not interpolated.
 #'
 #' @references
-#' \href{https://developer.here.com/documentation/destination-weather/dev_guide/topics/guide.html}{HERE Destination Weather API: Observation}
+#' \href{https://developer.here.com/documentation/destination-weather/api-reference-v3.html}{HERE Destination Weather API}
 #'
 #' @param poi \code{sf} object or character, Points of Interest (POIs) of geometry type \code{POINT} or location names (e.g. cities or regions).
-#' @param product character, weather product of the 'Destination Weather API'. Supported products: \code{"observation"}, \code{"forecast_hourly"}, \code{"forecast_astronomy"} and \code{"alerts"}.
+#' @param product character, weather product of the 'Destination Weather API'. Supported products: \code{"observation"}, \code{"forecastHourly"}, \code{"forecastAstronomy"} and \code{"alerts"}.
 #' @param url_only boolean, only return the generated URLs (\code{default = FALSE})?
 #'
 #' @return
@@ -25,57 +25,66 @@
 #' observation <- weather(poi = poi, product = "observation", url_only = TRUE)
 #'
 #' # Forecast
-#' forecast <- weather(poi = poi, product = "forecast_hourly", url_only = TRUE)
+#' forecast <- weather(poi = poi, product = "forecastHourly", url_only = TRUE)
 #'
 #' # Astronomy
-#' astronomy <- weather(poi = poi, product = "forecast_astronomy", url_only = TRUE)
+#' astronomy <- weather(poi = poi, product = "forecastAstronomy", url_only = TRUE)
 #'
 #' # Alerts
 #' alerts <- weather(poi = poi, product = "alerts", url_only = TRUE)
 weather <- function(poi, product = "observation", url_only = FALSE) {
+  UseMethod("weather", poi)
+}
+
+#' @export
+weather.character <- function(poi, product = "observation", url_only = FALSE) {
+  .check_character(poi)
+  query <- paste0(
+    "&q=",
+    curl::curl_escape(poi)
+  )
+  .weather.default(query, product, url_only)
+}
+
+#' @export
+weather.sf <- function(poi, product = "observation", url_only = FALSE) {
+  weather.sfc(sf::st_geometry(poi), product, url_only)
+}
+
+#' @export
+weather.sfc <- function(poi, product = "observation", url_only = FALSE) {
+  .check_points(poi)
+  poi <- sf::st_coordinates(
+    sf::st_transform(poi, 4326)
+  )
+  location <- paste0(
+    "&location=", poi[, 2], ",", poi[, 1]
+  )
+  .weather.default(location, product, url_only)
+}
+
+.weather.default <- function(poi, product = "observation", url_only = FALSE) {
   # Checks
   .check_weather_product(product)
   .check_boolean(url_only)
 
   # Add API key
   url <- .add_key(
-    url = "https://weather.ls.hereapi.com/weather/1.0/report.json?"
+    url = "https://weather.hereapi.com/v3/report?"
+  )
+
+  # Add formatted location or query parameter
+  url <- paste0(
+    url,
+    poi
   )
 
   # Add product
   url <- paste0(
     url,
-    "&product=",
+    "&products=",
     product
   )
-
-  # Check and preprocess location
-  # Character location (remove pipes)
-  if (is.character(poi)) {
-    .check_character(poi)
-    poi[poi == ""] <- NA
-    url <- paste0(
-      url,
-      "&name=",
-      curl::curl_escape(poi)
-    )
-    # sf POINTs
-  } else if ("sf" %in% class(poi)) {
-    .check_points(poi)
-    poi <- sf::st_coordinates(
-      sf::st_transform(poi, 4326)
-    )
-    poi <- paste0(
-      "&longitude=", poi[, 1], "&latitude=", poi[, 2]
-    )
-    url <- paste0(
-      url,
-      poi
-    )
-    # Not valid
-  } else {
-    stop("Invalid input for 'poi'.")
-  }
 
   # Return urls if chosen
   if (url_only) {
@@ -92,22 +101,19 @@ weather <- function(poi, product = "observation", url_only = FALSE) {
   }
 
   # Extract information
-  if (product == "observation") {
-    weather <- .extract_weather_observation(data)
-  } else if (product == "forecast_hourly") {
-    weather <- .extract_weather_forecast_hourly(data)
-  } else if (product == "forecast_astronomy") {
-    weather <- .extract_weather_forecast_astronomy(data)
-  } else if (product == "alerts") {
-    weather <- .extract_weather_alerts(data)
-  }
+  weather_data <- switch(product,
+    "observation" = .extract_weather_observation(data),
+    "forecastHourly" = .extract_weather_forecast_hourly(data),
+    "forecastAstronomy" = .extract_weather_forecast_astronomy(data),
+    "alerts" = .extract_weather_alerts(data)
+  )
 
   # Create sf, data.table, data.frame
-  rownames(weather) <- NULL
+  rownames(weather_data) <- NULL
   return(
     sf::st_set_crs(
       sf::st_as_sf(
-        as.data.frame(weather),
+        as.data.frame(weather_data),
         coords = c("lng", "lat")
       ), 4326
     )
@@ -115,114 +121,157 @@ weather <- function(poi, product = "observation", url_only = FALSE) {
 }
 
 .extract_weather_observation <- function(data) {
-  ids <- .get_ids(data)
   count <- 0
-  observation <- data.table::rbindlist(
+  observations <- data.table::rbindlist(
     lapply(data, function(con) {
       count <<- count + 1
       df <- jsonlite::fromJSON(con)
-      station <- data.table::data.table(
-        id = ids[count],
-        station = df$observations$location$city[1],
-        lng = df$observations$location$longitude[1],
-        lat = df$observations$location$latitude[1],
-        distance = df$observations$location$distance[1] * 1000,
-        timestamp = .parse_datetime(df$observations$location$observation[[1]]$utcTime),
-        state = df$observations$location$state[1],
-        country = df$observations$location$country[1]
+      data.table::rbindlist(
+        lapply(df$places$observations, function(result) {
+          res <- .parse_weather_results(result)
+          cbind(.parse_locations(result, count, nrow(res)), res)
+        }),
+        fill = TRUE
       )
-      obs <- df$observations$location$observation[[1]]
-      obs <- obs[, !names(obs) %in% c(
-        "skyDescription", "airDescription", "precipitationDesc",
-        "temperatureDesc", "iconName", "iconLink", "windDesc", "icon",
-        "country", "state", "city", "latitude", "longitude", "distance",
-        "utcTime", "elevation"
-      ), ]
-      obs[, c(4:9)] <- vapply(obs[, c(4:9)], as.numeric, numeric(1))
-      return(
-        cbind(station, obs)
-      )
-    })
+    }),
+    fill = TRUE
   )
-  return(observation)
+  return(observations)
 }
 
 .extract_weather_forecast_hourly <- function(data) {
-  ids <- .get_ids(data)
   count <- 0
-  dfs <- lapply(data, function(con) {
-    jsonlite::fromJSON(con)
-  })
+  forecasts <- list()
   forecast <- data.table::rbindlist(
-    lapply(dfs, function(df) {
+    lapply(data, function(con) {
       count <<- count + 1
-      station <- data.table::data.table(
-        id = ids[count],
-        station = df$hourlyForecasts$forecastLocation$city[1],
-        lng = df$hourlyForecasts$forecastLocation$longitude[1],
-        lat = df$hourlyForecasts$forecastLocation$latitude[1],
-        distance = df$hourlyForecasts$forecastLocation$distance[1] * 1000,
-        state = df$hourlyForecasts$forecastLocation$state[1],
-        country = df$hourlyForecasts$forecastLocation$country[1]
-      )
-    })
+      df <- jsonlite::fromJSON(con)
+      data.table::rbindlist(lapply(df$places$hourlyForecast, function(result) {
+        forecasts <<- append(forecasts, lapply(result$forecasts, .parse_weather_results))
+        .parse_locations(result, count, 1)
+      }), fill = TRUE)
+    }),
+    fill = TRUE
   )
-  forecast$forecast <- lapply(dfs, function(df) {
-    df$hourlyForecasts$forecastLocation$forecast
-  })
+  forecast <- cbind(forecast, forecasts)
   return(forecast)
 }
 
 .extract_weather_forecast_astronomy <- function(data) {
-  ids <- .get_ids(data)
   count <- 0
-  dfs <- lapply(data, function(con) {
-    jsonlite::fromJSON(con)
-  })
-  astronomy <- data.table::rbindlist(
-    lapply(dfs, function(df) {
+  forecasts <- list()
+  forecast <- data.table::rbindlist(
+    lapply(data, function(con) {
       count <<- count + 1
-      station <- data.table::data.table(
-        id = ids[count],
-        station = df$astronomy$city[1],
-        lng = df$astronomy$longitude[1],
-        lat = df$astronomy$latitude[1],
-        tz = df$astronomy$timezone[1],
-        state = df$astronomy$state[1],
-        country = df$astronomy$country[1]
-      )
-    })
+      df <- jsonlite::fromJSON(con)
+      data.table::rbindlist(lapply(df$places$astronomyForecasts, function(result) {
+        forecasts <<- append(forecasts, lapply(result$forecasts, .parse_astronomy_results))
+        .parse_locations(result, count, 1)
+      }), fill = TRUE)
+    }),
+    fill = TRUE
   )
-  astronomy$astronomy <- lapply(dfs, function(df) {
-    ast <- df$astronomy$astronomy
-    ast$date <- as.Date(.parse_datetime(ast$utcTime))
-    ast$utcTime <- NULL
-    ast
-  })
-  return(astronomy)
+  forecast <- cbind(forecast, forecasts)
+  return(forecast)
 }
 
 .extract_weather_alerts <- function(data) {
-  ids <- .get_ids(data)
   count <- 0
-  dfs <- lapply(data, function(con) {
-    jsonlite::fromJSON(con)
-  })
-  alerts <- data.table::rbindlist(
-    lapply(dfs, function(df) {
+  alerts <- list()
+  locations <- data.table::rbindlist(
+    lapply(data, function(con) {
       count <<- count + 1
-      station <- data.table::data.table(
-        id = ids[count],
-        station = df$alerts$city[1],
-        lng = df$alerts$longitude[1],
-        lat = df$alerts$latitude[1],
-        state = df$alerts$state[1],
-        country = df$alerts$country[1]
-      )
-    })
+      df <- jsonlite::fromJSON(con)
+      data.table::rbindlist(lapply(df$places$alerts, function(result) {
+        alerts <<- append(alerts, list(.parse_alert_results(result)))
+        .parse_locations(result, count, 1)
+      }), fill = TRUE)
+    }),
+    fill = TRUE
   )
-  alerts$alerts <- lapply(dfs, function(df) {
-    df$alerts$alerts
-  })
-  return(alerts)
+  return(cbind(locations, alerts))
+}
+
+.parse_locations <- function(df, req_id, max_rank_id) {
+  return(data.table::data.table(
+    id = req_id,
+    rank = seq_len(max_rank_id),
+    country_code = df$place$address$countryCode,
+    country = df$place$address$countryName,
+    state = df$place$address$state,
+    city = df$place$address$city,
+    distance = df$place$distance,
+    lng = df$place$location$lng,
+    lat = df$place$location$lat
+  ))
+}
+
+.parse_weather_results <- function(df) {
+  return(
+    data.table::data.table(
+      time = .parse_datetime_tz(df$time),
+      daylight = df$daylight,
+      description = df$description,
+      sky_info = df$skyInfo,
+      sky_desc = df$skyDesc,
+      temperature = df$temperature,
+      temperature_desc = df$temperatureDesc,
+      comfort = df$comfort,
+      high_temperature = .format_na_values(df$highTemperature),
+      low_temperature = .format_na_values(df$lowTemperature),
+      humidity = .format_na_values(df$humidity),
+      dew_point = df$dewPoint,
+      precipitation_probability = df$precipitationProbability,
+      rain_fall = df$rainFall,
+      wind_speed = df$windSpeed,
+      wind_direction = df$windDirection,
+      wind_descr = df$windDesc,
+      wind_descr_short = df$windDescShort,
+      uv_index = df$uvIndex,
+      uv_descr = df$uvDesc,
+      barometer_pressure = df$barometerPressure,
+      barometer_trend = df$barometerTrend,
+      age_minutes = df$ageMinutes,
+      active_alerts = df$activeAlerts
+    )
+  )
+}
+
+.parse_astronomy_results <- function(df) {
+  return(
+    data.table::data.table(
+      time = .parse_datetime_tz(df$time),
+      sun_rise = df$sunRise,
+      sun_set = df$sunSet,
+      moon_rise = df$moonRise,
+      moon_set = df$moonSet,
+      moon_phase = df$moonPhase,
+      moon_phase_description = df$moonPhaseDescription
+    )
+  )
+}
+
+.parse_alert_results <- function(df) {
+  template <- data.table::data.table(
+    time_segments = list(),
+    type = numeric(),
+    descriction = character()
+  )
+  return(
+    rbind(
+      template,
+      data.table::data.table(
+        time_segments = df$timeSegments,
+        type = ifelse(is.null(df$type), NA, df$type),
+        descriction = ifelse(is.null(df$description), NA, df$description)
+      )
+    )
+  )
+}
+
+.format_na_values <- function(col) {
+  if (is.null(col)) {
+    return(NULL)
+  }
+  as.numeric(gsub("\\*", "", col))
 }